+++ client/src/components/popularCard/PopularCard.tsx
@@ -6,7 +6,7 @@ import { movie } from '../../models/movies';
 import { useIsShow } from '../../hooks/useIsShow';
 import { useMatchGenreId } from '../../hooks/useMatchGenreId';
 import { Link } from 'react-router-dom';
-import { addToWatchlist } from '../../utils/addToWatchlist';
+import { addToWatchlistService } from '../../services/addToWatchlist.service';
 
 interface PopularCardProps {
   content: show | movie;
@@ -39,7 +39,10 @@ const PopularCard: React.FC<PopularCardProps> = ({ content }) => {
           className={'flex space-between align-items'}
           style={{ width: '100%' }}
         >
-          <button className={'plus'} onClick={() => addToWatchlist(content.id)}>
+          <button
+            className={'plus'}
+            onClick={() => addToWatchlistService(content.id)}
+          >
             +
           </button>
           <Link to={`details/${type}/${content.id}`}>
